Cleaned up (and moved) cfn_shp_to_vtk.py. Functions are now in separate modules. Added another script to create Tsurfs from CFM shapefile.
